+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ## [X.X.X]
 * Feature:
 	* Add readme pod shields
+	* Use RxSwift in example
+* Fixes:
+	* RxSwift compatibility
 
 ## [0.9.2]
 * Feature: