+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 * Fixes:
 	* RxSwift compatibility
 	* CommandCenter set when playing
+	* Don't play on interruption ended when playing from another app
 
 ## [0.9.2]
 * Feature: